Add API priority level and fix bug on removals
Sometimes we have more important messages we want to try before lower
priority messages before the deadline runs out, so now you can submit
requests with a priority level.

Lower numbers are higher piority (so events with 0, 40, 20, 100 will run
in order 0, 20, 40, 100) and if no priority is provided it defaults to
100. technically negative numbers also work as priorities too.

Also fixes a bug where we were using the incorrect variable for cleaning
up the speak history and some events just kept repeating over and over
and over again without being delted. Now, also due to the priority
sorting, we just iterate the speak events once and delete from the
original speak source without needing to double loop the list (and since
the error was in the second clean-up loop which is now gone, there's no
more errors, anywhere, ever again).

Also cleans up the logging/printing of events to be in the execution
management loop instead of in the speaking adapter itself, so there's
more consistent debug/log printing in more consistent places (also
removed the log-on-API-hit logger because it'll all get logged when
either triggered or skipped via the deadline breaching).