Fix bug where expired elements not being deleted
Bug introduced by us modifying the Event dict key directly (we shouldn't
be modifying these at all, they were originally Frozen, but we had to
remove the Frozen flag to enable sorting, then we started modifying it
and things broke).

Basically, (somehow?) after we deleted the event from the dictionary,
then we modified the event, it ended up not being deleted from the
dictionary? This had something to do with the hash keys and equality and
concurrent updates, but now it works agian and expired events aren't
infintie looping anymore.

Also adds a little more per-call debugging details about the current
length of the speaking queue being evaluated and the current element
being processed for the run.