Use self-sorting container for priority management
Much cleaner to just use a magic dict where the items are also sorting
themselves instead of running sorted() on the dict every time we loop
through it.

Much nicer than running N^2 sorts over the counter (but our N is low
because things are either running or quickly expired) and it maintains
priority levels on insert instead of re-calculating for every loop of
every element on every read.