Allow spoken events to have non-duplicate checks
For some spoken events it's nice to have "extra metadata" appended to
the end of a speech trigger, but we want the _core_ speech trigger to
still be counted for repeated speaking events.

So now there's a difference between the "core speech" string and an
optional "aux" string where only the "core speech" string is used for
detecting duplicate events to avoid, and any spoken event passing the
duplicate checks still gets the "aux" spoken anyway.